requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
pymysql
python-dotenv
dill
streamlit
gradio
setuptools
cryptography
